ATS-PRO/requirements.txt
#libraries need to be installed

streamlit
streamlit_extras
google-generativeai
python-dotenv
PyPDF2
Pillow
